Moved gridImageProperties into helpers directory.
